+++ src/main/java/kitchenpos/dao/OrderLineItemDao.java
@@ -1,6 +1,6 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.OrderLineItemRequest;
+import kitchenpos.table.dto.OrderLineItemRequest;
 
 import java.util.List;
 import java.util.Optional;
